+++ app/src/main/java/com/amberleesmith/beautifulbulldog/NewBulldogActivity.java
@@ -1,0 +1,13 @@
+package com.amberleesmith.beautifulbulldog;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class NewBulldogActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_new_bulldog);
+    }
+}
